+++ NEWS
@@ -1,3 +1,7 @@
+05.02.2021 Version 0.0.27
+RestrictTo field support added.
+Everyone option added to permission dropdown.
+
 04.02.2021 Version 0.0.26
 Settings permission feature added for forms.
