+++ models/collectionsModel.js
@@ -86,8 +86,7 @@ const collectionsSchema = new mongoose.Schema(
       type: 'Mixed'
     },
     restrictTo: {
-      type: 'Mixed',
-      default: { role: ['admin'] }
+      type: 'Mixed'
     },
     owner: {
       type: mongoose.Schema.ObjectId,
